Modularize CLI drive orchestration helpers
