--- crates/engine/src/local_copy/executor/file.rs
@@ -1,1442 +1,0 @@
-//! Regular file copy routines and helpers.
-
-use std::ffi::{OsStr, OsString};
-use std::fs;
-use std::io::{self, Read, Seek, SeekFrom, Write};
-use std::num::NonZeroU8;
-use std::path::{Path, PathBuf};
-use std::process;
-use std::sync::atomic::Ordering as AtomicOrdering;
-use std::time::{Duration, Instant, SystemTime};
-
-#[cfg(unix)]
-use rustix::{
-    fd::AsFd,
-    fs::{FallocateFlags, fallocate},
-    io::Errno,
-};
-
-use crate::delta::{DeltaSignatureIndex, SignatureLayoutParams, calculate_signature_layout};
-use crate::signature::{SignatureAlgorithm, SignatureError, generate_file_signature};
-
-use crate::local_copy::overrides::create_hard_link;
-use crate::local_copy::{
-    CopyContext, CreatedEntryKind, DeferredUpdate, FinalizeMetadataParams, LocalCopyAction,
-    LocalCopyArgumentError, LocalCopyError, LocalCopyMetadata, LocalCopyRecord, ReferenceDecision,
-    ReferenceQuery, create_symlink, find_reference_action, map_metadata_error,
-    remove_source_entry_if_requested,
-};
-
-#[cfg(feature = "acl")]
-use crate::local_copy::sync_acls_if_requested;
-#[cfg(feature = "xattr")]
-use crate::local_copy::sync_xattrs_if_requested;
-
-use rsync_checksums::strong::{Md4, Md5, Xxh3, Xxh3_128, Xxh64};
-use rsync_meta::{MetadataOptions, apply_file_metadata_with_options};
-use rsync_protocol::ProtocolVersion;
-
-use super::super::{COPY_BUFFER_SIZE, CROSS_DEVICE_ERROR_CODE, NEXT_TEMP_FILE_ID};
-
-pub(crate) fn maybe_preallocate_destination(
-    file: &mut fs::File,
-    path: &Path,
-    total_len: u64,
-    existing_bytes: u64,
-    enabled: bool,
-) -> Result<(), LocalCopyError> {
-    if !enabled || total_len == 0 || total_len <= existing_bytes {
-        return Ok(());
-    }
-
-    preallocate_destination_file(file, path, total_len)
-}
-
-fn preallocate_destination_file(
-    file: &mut fs::File,
-    path: &Path,
-    total_len: u64,
-) -> Result<(), LocalCopyError> {
-    #[cfg(unix)]
-    {
-        if total_len == 0 {
-            return Ok(());
-        }
-
-        if total_len > i64::MAX as u64 {
-            return Err(LocalCopyError::io(
-                "preallocate destination file",
-                path.to_path_buf(),
-                io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    "preallocation size exceeds platform limit",
-                ),
-            ));
-        }
-
-        let fd = file.as_fd();
-        match fallocate(fd, FallocateFlags::empty(), 0, total_len) {
-            Ok(()) => Ok(()),
-            Err(Errno::OPNOTSUPP | Errno::NOSYS | Errno::INVAL) => {
-                file.set_len(total_len).map_err(|error| {
-                    LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
-                })
-            }
-            Err(errno) => Err(LocalCopyError::io(
-                "preallocate destination file",
-                path.to_path_buf(),
-                io::Error::from_raw_os_error(errno.raw_os_error()),
-            )),
-        }
-    }
-
-    #[cfg(not(unix))]
-    {
-        if total_len == 0 {
-            return Ok(());
-        }
-
-        file.set_len(total_len).map_err(|error| {
-            LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
-        })
-    }
-}
-
-pub(crate) fn copy_file(
-    context: &mut CopyContext,
-    source: &Path,
-    destination: &Path,
-    metadata: &fs::Metadata,
-    relative: Option<&Path>,
-) -> Result<(), LocalCopyError> {
-    context.enforce_timeout()?;
-    let metadata_options = context.metadata_options();
-    let mode = context.mode();
-    let file_type = metadata.file_type();
-    #[cfg(feature = "xattr")]
-    let preserve_xattrs = context.xattrs_enabled();
-    #[cfg(feature = "acl")]
-    let preserve_acls = context.acls_enabled();
-    let record_path = relative
-        .map(Path::to_path_buf)
-        .or_else(|| source.file_name().map(PathBuf::from))
-        .unwrap_or_else(|| {
-            destination
-                .file_name()
-                .map(PathBuf::from)
-                .unwrap_or_default()
-        });
-    let file_size = metadata.len();
-    context.summary_mut().record_regular_file_total();
-    context.summary_mut().record_total_bytes(file_size);
-
-    if let Some(min_limit) = context.min_file_size_limit() {
-        if file_size < min_limit {
-            return Ok(());
-        }
-    }
-
-    if let Some(max_limit) = context.max_file_size_limit() {
-        if file_size > max_limit {
-            return Ok(());
-        }
-    }
-    if let Some(parent) = destination.parent() {
-        context.prepare_parent_directory(parent)?;
-    }
-
-    let existing_metadata = match fs::symlink_metadata(destination) {
-        Ok(existing) => Some(existing),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => None,
-        Err(error) => {
-            return Err(LocalCopyError::io(
-                "inspect existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-    };
-
-    if let Some(existing) = &existing_metadata {
-        if existing.file_type().is_dir() {
-            return Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::ReplaceDirectoryWithFile,
-            ));
-        }
-    }
-
-    let destination_previously_existed = existing_metadata.is_some();
-
-    if mode.is_dry_run() {
-        if context.update_enabled() {
-            if let Some(existing) = existing_metadata.as_ref() {
-                if destination_is_newer(metadata, existing) {
-                    context.summary_mut().record_regular_file_skipped_newer();
-                    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                    let total_bytes = Some(metadata_snapshot.len());
-                    context.record(LocalCopyRecord::new(
-                        record_path.clone(),
-                        LocalCopyAction::SkippedNewerDestination,
-                        0,
-                        total_bytes,
-                        Duration::default(),
-                        Some(metadata_snapshot),
-                    ));
-                    return Ok(());
-                }
-            }
-        }
-
-        if context.ignore_existing_enabled() && existing_metadata.is_some() {
-            context.summary_mut().record_regular_file_ignored_existing();
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::SkippedExisting,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-            return Ok(());
-        }
-
-        let mut reader = fs::File::open(source)
-            .map_err(|error| LocalCopyError::io("open source file", source.to_path_buf(), error))?;
-
-        let append_mode = determine_append_mode(
-            context.append_enabled(),
-            context.append_verify_enabled(),
-            &mut reader,
-            source,
-            destination,
-            existing_metadata.as_ref(),
-            file_size,
-        )?;
-        let append_offset = match append_mode {
-            AppendMode::Append(offset) => offset,
-            AppendMode::Disabled => 0,
-        };
-        let bytes_transferred = file_size.saturating_sub(append_offset);
-
-        context
-            .summary_mut()
-            .record_file(file_size, bytes_transferred, None);
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(
-            LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::DataCopied,
-                bytes_transferred,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            )
-            .with_creation(!destination_previously_existed),
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-
-    if context.update_enabled() {
-        if let Some(existing) = existing_metadata.as_ref() {
-            if destination_is_newer(metadata, existing) {
-                context.summary_mut().record_regular_file_skipped_newer();
-                context.record_hard_link(metadata, destination);
-                let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                let total_bytes = Some(metadata_snapshot.len());
-                context.record(LocalCopyRecord::new(
-                    record_path.clone(),
-                    LocalCopyAction::SkippedNewerDestination,
-                    0,
-                    total_bytes,
-                    Duration::default(),
-                    Some(metadata_snapshot),
-                ));
-                return Ok(());
-            }
-        }
-    }
-
-    if context.ignore_existing_enabled() && existing_metadata.is_some() {
-        context.summary_mut().record_regular_file_ignored_existing();
-        context.record_hard_link(metadata, destination);
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::SkippedExisting,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        return Ok(());
-    }
-
-    let use_sparse_writes = context.sparse_enabled();
-    let partial_enabled = context.partial_enabled();
-    let inplace_enabled = context.inplace_enabled();
-    let checksum_enabled = context.checksum_enabled();
-    let size_only_enabled = context.size_only_enabled();
-    let append_allowed = context.append_enabled();
-    let append_verify = context.append_verify_enabled();
-    let whole_file_enabled = context.whole_file_enabled();
-    let compress_enabled = context.should_compress(record_path.as_path());
-    let relative_for_link = relative.unwrap_or(record_path.as_path());
-
-    if let Some(existing) = existing_metadata.as_ref() {
-        context.backup_existing_entry(destination, relative, existing.file_type())?;
-    }
-
-    if let Some(link_target) = context.link_dest_target(
-        relative_for_link,
-        source,
-        metadata,
-        &metadata_options,
-        size_only_enabled,
-        checksum_enabled,
-    )? {
-        let mut attempted_commit = false;
-        loop {
-            match fs::hard_link(&link_target, destination) {
-                Ok(()) => break,
-                Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                    remove_existing_destination(destination)?;
-                    fs::hard_link(&link_target, destination).map_err(|link_error| {
-                        LocalCopyError::io(
-                            "create hard link",
-                            destination.to_path_buf(),
-                            link_error,
-                        )
-                    })?;
-                    break;
-                }
-                Err(error)
-                    if error.kind() == io::ErrorKind::NotFound
-                        && context.delay_updates_enabled()
-                        && !attempted_commit =>
-                {
-                    context.commit_deferred_update_for(&link_target)?;
-                    attempted_commit = true;
-                    continue;
-                }
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "create hard link",
-                        destination.to_path_buf(),
-                        error,
-                    ));
-                }
-            }
-        }
-
-        context.record_hard_link(metadata, destination);
-        context.summary_mut().record_hard_link();
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::HardLink,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        context.register_created_path(
-            destination,
-            CreatedEntryKind::HardLink,
-            destination_previously_existed,
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-    let mut copy_source_override: Option<PathBuf> = None;
-
-    if let Some(existing_target) = context.existing_hard_link_target(metadata) {
-        let mut attempted_commit = false;
-        loop {
-            match create_hard_link(&existing_target, destination) {
-                Ok(()) => break,
-                Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                    remove_existing_destination(destination)?;
-                    create_hard_link(&existing_target, destination).map_err(|link_error| {
-                        LocalCopyError::io(
-                            "create hard link",
-                            destination.to_path_buf(),
-                            link_error,
-                        )
-                    })?;
-                    break;
-                }
-                Err(error)
-                    if error.kind() == io::ErrorKind::NotFound
-                        && context.delay_updates_enabled()
-                        && !attempted_commit =>
-                {
-                    context.commit_deferred_update_for(&existing_target)?;
-                    attempted_commit = true;
-                    continue;
-                }
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "create hard link",
-                        destination.to_path_buf(),
-                        error,
-                    ));
-                }
-            }
-        }
-
-        context.record_hard_link(metadata, destination);
-        context.summary_mut().record_hard_link();
-        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-        let total_bytes = Some(metadata_snapshot.len());
-        context.record(LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::HardLink,
-            0,
-            total_bytes,
-            Duration::default(),
-            Some(metadata_snapshot),
-        ));
-        context.register_created_path(
-            destination,
-            CreatedEntryKind::HardLink,
-            destination_previously_existed,
-        );
-        remove_source_entry_if_requested(context, source, Some(record_path.as_path()), file_type)?;
-        return Ok(());
-    }
-
-    if !context.reference_directories().is_empty() && !record_path.as_os_str().is_empty() {
-        if let Some(decision) = find_reference_action(
-            context,
-            ReferenceQuery {
-                destination,
-                relative: record_path.as_path(),
-                source,
-                metadata,
-                metadata_options: &metadata_options,
-                size_only: size_only_enabled,
-                checksum: checksum_enabled,
-            },
-        )? {
-            match decision {
-                ReferenceDecision::Skip => {
-                    context.summary_mut().record_regular_file_matched();
-                    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                    let total_bytes = Some(metadata_snapshot.len());
-                    context.record(LocalCopyRecord::new(
-                        record_path.clone(),
-                        LocalCopyAction::MetadataReused,
-                        0,
-                        total_bytes,
-                        Duration::default(),
-                        Some(metadata_snapshot),
-                    ));
-                    context.register_progress();
-                    remove_source_entry_if_requested(
-                        context,
-                        source,
-                        Some(record_path.as_path()),
-                        file_type,
-                    )?;
-                    return Ok(());
-                }
-                ReferenceDecision::Copy(path) => {
-                    copy_source_override = Some(path);
-                }
-                ReferenceDecision::Link(path) => {
-                    if existing_metadata.is_some() {
-                        remove_existing_destination(destination)?;
-                    }
-
-                    let link_result = create_hard_link(&path, destination);
-                    let mut degrade_to_copy = false;
-                    match link_result {
-                        Ok(()) => {}
-                        Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                            remove_existing_destination(destination)?;
-                            create_hard_link(&path, destination).map_err(|link_error| {
-                                LocalCopyError::io(
-                                    "create hard link",
-                                    destination.to_path_buf(),
-                                    link_error,
-                                )
-                            })?;
-                        }
-                        Err(error)
-                            if matches!(
-                                error.raw_os_error(),
-                                Some(code) if code == CROSS_DEVICE_ERROR_CODE
-                            ) =>
-                        {
-                            degrade_to_copy = true;
-                        }
-                        Err(error) => {
-                            return Err(LocalCopyError::io(
-                                "create hard link",
-                                destination.to_path_buf(),
-                                error,
-                            ));
-                        }
-                    }
-
-                    if degrade_to_copy {
-                        copy_source_override = Some(path);
-                    } else if copy_source_override.is_none() {
-                        apply_file_metadata_with_options(
-                            destination,
-                            metadata,
-                            metadata_options.clone(),
-                        )
-                        .map_err(map_metadata_error)?;
-                        #[cfg(feature = "xattr")]
-                        sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-                        #[cfg(feature = "acl")]
-                        sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-                        context.record_hard_link(metadata, destination);
-                        context.summary_mut().record_hard_link();
-                        let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-                        let total_bytes = Some(metadata_snapshot.len());
-                        context.record(LocalCopyRecord::new(
-                            record_path.clone(),
-                            LocalCopyAction::HardLink,
-                            0,
-                            total_bytes,
-                            Duration::default(),
-                            Some(metadata_snapshot),
-                        ));
-                        context.register_created_path(
-                            destination,
-                            CreatedEntryKind::HardLink,
-                            destination_previously_existed,
-                        );
-                        context.register_progress();
-                        remove_source_entry_if_requested(
-                            context,
-                            source,
-                            Some(record_path.as_path()),
-                            file_type,
-                        )?;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-    }
-
-    if let Some(existing) = existing_metadata.as_ref() {
-        if should_skip_copy(CopyComparison {
-            source_path: source,
-            source: metadata,
-            destination_path: destination,
-            destination: existing,
-            options: &metadata_options,
-            size_only: size_only_enabled,
-            checksum: checksum_enabled,
-            checksum_algorithm: context.options().checksum_algorithm(),
-            modify_window: context.options().modify_window(),
-        }) {
-            apply_file_metadata_with_options(destination, metadata, metadata_options.clone())
-                .map_err(map_metadata_error)?;
-            #[cfg(feature = "xattr")]
-            sync_xattrs_if_requested(preserve_xattrs, mode, source, destination, true)?;
-            #[cfg(feature = "acl")]
-            sync_acls_if_requested(preserve_acls, mode, source, destination, true)?;
-            context.record_hard_link(metadata, destination);
-            context.summary_mut().record_regular_file_matched();
-            let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-            let total_bytes = Some(metadata_snapshot.len());
-            context.record(LocalCopyRecord::new(
-                record_path.clone(),
-                LocalCopyAction::MetadataReused,
-                0,
-                total_bytes,
-                Duration::default(),
-                Some(metadata_snapshot),
-            ));
-            return Ok(());
-        }
-    }
-
-    let mut reader = fs::File::open(source)
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let append_mode = determine_append_mode(
-        append_allowed,
-        append_verify,
-        &mut reader,
-        source,
-        destination,
-        existing_metadata.as_ref(),
-        file_size,
-    )?;
-    let append_offset = match append_mode {
-        AppendMode::Append(offset) => offset,
-        AppendMode::Disabled => 0,
-    };
-    reader
-        .seek(SeekFrom::Start(append_offset))
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let delta_signature = if append_offset == 0 && !whole_file_enabled && !inplace_enabled {
-        match existing_metadata.as_ref() {
-            Some(existing) if existing.is_file() => build_delta_signature(destination, existing)?,
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    let copy_source = copy_source_override.as_deref().unwrap_or(source);
-    let mut reader = fs::File::open(copy_source)
-        .map_err(|error| LocalCopyError::io("copy file", copy_source.to_path_buf(), error))?;
-    if append_offset > 0 {
-        reader
-            .seek(SeekFrom::Start(append_offset))
-            .map_err(|error| LocalCopyError::io("copy file", copy_source.to_path_buf(), error))?;
-    }
-    let mut guard = None;
-    let mut staging_path: Option<PathBuf> = None;
-
-    let mut writer = if append_offset > 0 {
-        let mut file = fs::OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(false)
-            .open(destination)
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?;
-        file.seek(SeekFrom::Start(append_offset))
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?;
-        file
-    } else if inplace_enabled {
-        fs::OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(true)
-            .open(destination)
-            .map_err(|error| LocalCopyError::io("copy file", destination.to_path_buf(), error))?
-    } else {
-        let (new_guard, file) = DestinationWriteGuard::new(
-            destination,
-            partial_enabled,
-            context.partial_directory_path(),
-            context.temp_directory_path(),
-        )?;
-        staging_path = Some(new_guard.staging_path().to_path_buf());
-        guard = Some(new_guard);
-        file
-    };
-    let preallocate_target = guard
-        .as_ref()
-        .map(|existing_guard| existing_guard.staging_path())
-        .unwrap_or(destination);
-    maybe_preallocate_destination(
-        &mut writer,
-        preallocate_target,
-        file_size,
-        append_offset,
-        context.preallocate_enabled(),
-    )?;
-    let mut buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    let start = Instant::now();
-
-    let copy_result = context.copy_file_contents(
-        &mut reader,
-        &mut writer,
-        &mut buffer,
-        use_sparse_writes,
-        compress_enabled,
-        source,
-        destination,
-        record_path.as_path(),
-        delta_signature.as_ref(),
-        file_size,
-        append_offset,
-        start,
-    );
-
-    drop(writer);
-
-    let staging_path_for_links = guard
-        .as_ref()
-        .map(|existing_guard| existing_guard.staging_path().to_path_buf())
-        .or_else(|| staging_path.take());
-    let delay_updates_enabled = context.delay_updates_enabled();
-
-    let outcome = match copy_result {
-        Ok(outcome) => {
-            if let Err(timeout_error) = context.enforce_timeout() {
-                if let Some(guard) = guard.take() {
-                    guard.discard();
-                }
-
-                if existing_metadata.is_none() {
-                    remove_incomplete_destination(destination);
-                }
-
-                return Err(timeout_error);
-            }
-            outcome
-        }
-        Err(error) => {
-            if let Some(guard) = guard.take() {
-                guard.discard();
-            }
-
-            if existing_metadata.is_none() {
-                remove_incomplete_destination(destination);
-            }
-
-            return Err(error);
-        }
-    };
-
-    context.register_created_path(
-        destination,
-        CreatedEntryKind::File,
-        destination_previously_existed,
-    );
-
-    let hard_link_path = if delay_updates_enabled {
-        staging_path_for_links.as_deref().unwrap_or(destination)
-    } else {
-        destination
-    };
-    context.record_hard_link(metadata, hard_link_path);
-    let elapsed = start.elapsed();
-    let compressed_bytes = outcome.compressed_bytes();
-    context
-        .summary_mut()
-        .record_file(file_size, outcome.literal_bytes(), compressed_bytes);
-    context.summary_mut().record_elapsed(elapsed);
-    let metadata_snapshot = LocalCopyMetadata::from_metadata(metadata, None);
-    let total_bytes = Some(metadata_snapshot.len());
-    context.record(
-        LocalCopyRecord::new(
-            record_path.clone(),
-            LocalCopyAction::DataCopied,
-            outcome.literal_bytes(),
-            total_bytes,
-            elapsed,
-            Some(metadata_snapshot),
-        )
-        .with_creation(!destination_previously_existed),
-    );
-
-    if let Err(timeout_error) = context.enforce_timeout() {
-        if existing_metadata.is_none() {
-            remove_incomplete_destination(destination);
-        }
-
-        return Err(timeout_error);
-    }
-
-    let relative_for_removal = Some(record_path.clone());
-    if let Some(guard) = guard {
-        if delay_updates_enabled {
-            let destination_path = guard.final_path().to_path_buf();
-            let update = DeferredUpdate::new(
-                guard,
-                metadata.clone(),
-                metadata_options.clone(),
-                mode,
-                source.to_path_buf(),
-                relative_for_removal.clone(),
-                destination_path,
-                file_type,
-                destination_previously_existed,
-                #[cfg(feature = "xattr")]
-                preserve_xattrs,
-                #[cfg(feature = "acl")]
-                preserve_acls,
-            );
-            context.register_deferred_update(update);
-        } else {
-            let destination_path = guard.final_path().to_path_buf();
-            guard.commit()?;
-            context.apply_metadata_and_finalize(
-                destination_path.as_path(),
-                FinalizeMetadataParams::new(
-                    metadata,
-                    metadata_options.clone(),
-                    mode,
-                    source,
-                    relative_for_removal.as_deref(),
-                    file_type,
-                    destination_previously_existed,
-                    #[cfg(feature = "xattr")]
-                    preserve_xattrs,
-                    #[cfg(feature = "acl")]
-                    preserve_acls,
-                ),
-            )?;
-        }
-    } else {
-        context.apply_metadata_and_finalize(
-            destination,
-            FinalizeMetadataParams::new(
-                metadata,
-                metadata_options,
-                mode,
-                source,
-                relative_for_removal.as_deref(),
-                file_type,
-                destination_previously_existed,
-                #[cfg(feature = "xattr")]
-                preserve_xattrs,
-                #[cfg(feature = "acl")]
-                preserve_acls,
-            ),
-        )?;
-    }
-
-    Ok(())
-}
-
-pub(crate) enum AppendMode {
-    Disabled,
-    Append(u64),
-}
-
-pub(crate) fn determine_append_mode(
-    append_allowed: bool,
-    append_verify: bool,
-    reader: &mut fs::File,
-    source: &Path,
-    destination: &Path,
-    existing_metadata: Option<&fs::Metadata>,
-    file_size: u64,
-) -> Result<AppendMode, LocalCopyError> {
-    if !append_allowed {
-        return Ok(AppendMode::Disabled);
-    }
-
-    let existing = match existing_metadata {
-        Some(meta) if meta.is_file() => meta,
-        _ => return Ok(AppendMode::Disabled),
-    };
-
-    let existing_len = existing.len();
-    if existing_len == 0 || existing_len >= file_size {
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        return Ok(AppendMode::Disabled);
-    }
-
-    if append_verify {
-        let matches = verify_append_prefix(reader, source, destination, existing_len)?;
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        if !matches {
-            return Ok(AppendMode::Disabled);
-        }
-    } else {
-        reader
-            .seek(SeekFrom::Start(0))
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    }
-
-    Ok(AppendMode::Append(existing_len))
-}
-
-fn verify_append_prefix(
-    reader: &mut fs::File,
-    source: &Path,
-    destination: &Path,
-    existing_len: u64,
-) -> Result<bool, LocalCopyError> {
-    reader
-        .seek(SeekFrom::Start(0))
-        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-    let mut destination_file = fs::File::open(destination).map_err(|error| {
-        LocalCopyError::io(
-            "read existing destination",
-            destination.to_path_buf(),
-            error,
-        )
-    })?;
-    let mut remaining = existing_len;
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    while remaining > 0 {
-        let chunk = remaining.min(COPY_BUFFER_SIZE as u64) as usize;
-        let source_read = reader
-            .read(&mut source_buffer[..chunk])
-            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
-        let destination_read = destination_file
-            .read(&mut destination_buffer[..chunk])
-            .map_err(|error| {
-                LocalCopyError::io(
-                    "read existing destination",
-                    destination.to_path_buf(),
-                    error,
-                )
-            })?;
-
-        if source_read == 0 || destination_read == 0 || source_read != destination_read {
-            return Ok(false);
-        }
-
-        if source_buffer[..source_read] != destination_buffer[..destination_read] {
-            return Ok(false);
-        }
-
-        remaining = remaining.saturating_sub(source_read as u64);
-    }
-
-    Ok(true)
-}
-
-pub(crate) fn partial_destination_path(destination: &Path) -> PathBuf {
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_string_lossy().to_string())
-        .unwrap_or_else(|| "partial".to_string());
-    let partial_name = format!(".rsync-partial-{}", file_name);
-    destination.with_file_name(partial_name)
-}
-
-pub(crate) fn partial_directory_destination_path(
-    destination: &Path,
-    partial_dir: &Path,
-) -> Result<PathBuf, LocalCopyError> {
-    let base_dir = if partial_dir.is_absolute() {
-        partial_dir.to_path_buf()
-    } else {
-        let parent = destination
-            .parent()
-            .map_or_else(|| PathBuf::from("."), Path::to_path_buf);
-        parent.join(partial_dir)
-    };
-    fs::create_dir_all(&base_dir)
-        .map_err(|error| LocalCopyError::io("create partial directory", base_dir.clone(), error))?;
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_os_string())
-        .unwrap_or_else(|| OsStr::new("partial").to_os_string());
-    Ok(base_dir.join(file_name))
-}
-
-pub(crate) fn compute_backup_path(
-    destination_root: &Path,
-    destination: &Path,
-    relative: Option<&Path>,
-    backup_dir: Option<&Path>,
-    suffix: &OsStr,
-) -> PathBuf {
-    let relative_path = if let Some(rel) = relative {
-        rel.to_path_buf()
-    } else if let Ok(stripped) = destination.strip_prefix(destination_root) {
-        if stripped.as_os_str().is_empty() {
-            destination
-                .file_name()
-                .map(PathBuf::from)
-                .unwrap_or_else(|| PathBuf::from(destination))
-        } else {
-            stripped.to_path_buf()
-        }
-    } else if let Some(name) = destination.file_name() {
-        PathBuf::from(name)
-    } else {
-        PathBuf::from(destination)
-    };
-
-    let mut backup_name = relative_path
-        .file_name()
-        .map(|name| name.to_os_string())
-        .unwrap_or_else(|| OsString::from("backup"));
-    if !suffix.is_empty() {
-        backup_name.push(suffix);
-    }
-
-    let mut base = if let Some(dir) = backup_dir {
-        let mut base = if dir.is_absolute() {
-            dir.to_path_buf()
-        } else {
-            destination_root.join(dir)
-        };
-        if let Some(parent) = relative_path.parent() {
-            if !parent.as_os_str().is_empty() {
-                base = base.join(parent);
-            }
-        }
-        base
-    } else {
-        destination
-            .parent()
-            .map(Path::to_path_buf)
-            .unwrap_or_else(|| PathBuf::from("."))
-    };
-
-    base.push(backup_name);
-    base
-}
-
-pub(crate) fn copy_entry_to_backup(
-    source: &Path,
-    backup_path: &Path,
-    file_type: fs::FileType,
-) -> Result<(), LocalCopyError> {
-    if file_type.is_file() {
-        fs::copy(source, backup_path).map_err(|error| {
-            LocalCopyError::io("create backup", backup_path.to_path_buf(), error)
-        })?;
-    } else if file_type.is_symlink() {
-        let target = fs::read_link(source).map_err(|error| {
-            LocalCopyError::io("read symbolic link", source.to_path_buf(), error)
-        })?;
-        create_symlink(&target, source, backup_path).map_err(|error| {
-            LocalCopyError::io("create symbolic link", backup_path.to_path_buf(), error)
-        })?;
-    }
-    Ok(())
-}
-
-pub(crate) fn remove_existing_destination(path: &Path) -> Result<(), LocalCopyError> {
-    match fs::remove_file(path) {
-        Ok(()) => Ok(()),
-        Err(error) if error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(error) => Err(LocalCopyError::io(
-            "remove existing destination",
-            path.to_path_buf(),
-            error,
-        )),
-    }
-}
-
-pub(crate) fn temporary_destination_path(
-    destination: &Path,
-    unique: usize,
-    temp_dir: Option<&Path>,
-) -> PathBuf {
-    let file_name = destination
-        .file_name()
-        .map(|name| name.to_string_lossy().to_string())
-        .unwrap_or_else(|| "temp".to_string());
-    let temp_name = format!(".rsync-tmp-{file_name}-{}-{}", process::id(), unique);
-    match temp_dir {
-        Some(dir) => dir.join(temp_name),
-        None => destination.with_file_name(temp_name),
-    }
-}
-
-pub(crate) struct DestinationWriteGuard {
-    final_path: PathBuf,
-    temp_path: PathBuf,
-    preserve_on_error: bool,
-    committed: bool,
-}
-
-impl DestinationWriteGuard {
-    pub(crate) fn new(
-        destination: &Path,
-        partial: bool,
-        partial_dir: Option<&Path>,
-        temp_dir: Option<&Path>,
-    ) -> Result<(Self, fs::File), LocalCopyError> {
-        if partial {
-            let temp_path = if let Some(dir) = partial_dir {
-                partial_directory_destination_path(destination, dir)?
-            } else {
-                partial_destination_path(destination)
-            };
-            if let Err(error) = fs::remove_file(&temp_path) {
-                if error.kind() != io::ErrorKind::NotFound {
-                    return Err(LocalCopyError::io(
-                        "remove existing partial file",
-                        temp_path.clone(),
-                        error,
-                    ));
-                }
-            }
-            let file = fs::OpenOptions::new()
-                .create(true)
-                .write(true)
-                .truncate(true)
-                .open(&temp_path)
-                .map_err(|error| LocalCopyError::io("copy file", temp_path.clone(), error))?;
-            Ok((
-                Self {
-                    final_path: destination.to_path_buf(),
-                    temp_path,
-                    preserve_on_error: true,
-                    committed: false,
-                },
-                file,
-            ))
-        } else {
-            loop {
-                let unique = NEXT_TEMP_FILE_ID.fetch_add(1, AtomicOrdering::Relaxed);
-                let temp_path = temporary_destination_path(destination, unique, temp_dir);
-                match fs::OpenOptions::new()
-                    .write(true)
-                    .create_new(true)
-                    .open(&temp_path)
-                {
-                    Ok(file) => {
-                        return Ok((
-                            Self {
-                                final_path: destination.to_path_buf(),
-                                temp_path,
-                                preserve_on_error: false,
-                                committed: false,
-                            },
-                            file,
-                        ));
-                    }
-                    Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                        continue;
-                    }
-                    Err(error) => {
-                        return Err(LocalCopyError::io("copy file", temp_path.clone(), error));
-                    }
-                }
-            }
-        }
-    }
-
-    pub(crate) fn staging_path(&self) -> &Path {
-        &self.temp_path
-    }
-
-    pub(crate) fn commit(mut self) -> Result<(), LocalCopyError> {
-        match fs::rename(&self.temp_path, &self.final_path) {
-            Ok(()) => {}
-            Err(error) if error.kind() == io::ErrorKind::AlreadyExists => {
-                remove_existing_destination(&self.final_path)?;
-                fs::rename(&self.temp_path, &self.final_path).map_err(|rename_error| {
-                    LocalCopyError::io(self.finalise_action(), self.temp_path.clone(), rename_error)
-                })?;
-            }
-            Err(error) if error.kind() == io::ErrorKind::CrossesDevices => {
-                fs::copy(&self.temp_path, &self.final_path).map_err(|copy_error| {
-                    LocalCopyError::io(self.finalise_action(), self.final_path.clone(), copy_error)
-                })?;
-                fs::remove_file(&self.temp_path).map_err(|remove_error| {
-                    LocalCopyError::io(self.finalise_action(), self.temp_path.clone(), remove_error)
-                })?;
-            }
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    self.finalise_action(),
-                    self.temp_path.clone(),
-                    error,
-                ));
-            }
-        }
-        self.committed = true;
-        Ok(())
-    }
-
-    pub(crate) fn final_path(&self) -> &Path {
-        &self.final_path
-    }
-
-    fn discard(mut self) {
-        if self.preserve_on_error {
-            self.committed = true;
-            return;
-        }
-
-        if let Err(error) = fs::remove_file(&self.temp_path) {
-            if error.kind() != io::ErrorKind::NotFound {
-                // Best-effort cleanup: the file may have been removed concurrently.
-            }
-        }
-
-        self.committed = true;
-    }
-
-    fn finalise_action(&self) -> &'static str {
-        if self.preserve_on_error {
-            "finalise partial file"
-        } else {
-            "finalise temporary file"
-        }
-    }
-}
-
-impl Drop for DestinationWriteGuard {
-    fn drop(&mut self) {
-        if !self.committed && !self.preserve_on_error {
-            let _ = fs::remove_file(&self.temp_path);
-        }
-    }
-}
-
-fn remove_incomplete_destination(destination: &Path) {
-    if let Err(error) = fs::remove_file(destination) {
-        if error.kind() != io::ErrorKind::NotFound {
-            // Preserve the original error from the transfer attempt.
-        }
-    }
-}
-
-pub(crate) fn write_sparse_chunk(
-    writer: &mut fs::File,
-    chunk: &[u8],
-    destination: &Path,
-) -> Result<usize, LocalCopyError> {
-    let mut index = 0usize;
-    let mut written = 0usize;
-
-    while index < chunk.len() {
-        if chunk[index] == 0 {
-            let start = index;
-            while index < chunk.len() && chunk[index] == 0 {
-                index += 1;
-            }
-            let span = index - start;
-            if span > 0 {
-                writer
-                    .seek(SeekFrom::Current(span as i64))
-                    .map_err(|error| {
-                        LocalCopyError::io(
-                            "seek in destination file",
-                            destination.to_path_buf(),
-                            error,
-                        )
-                    })?;
-            }
-        } else {
-            let start = index;
-            while index < chunk.len() && chunk[index] != 0 {
-                index += 1;
-            }
-            writer.write_all(&chunk[start..index]).map_err(|error| {
-                LocalCopyError::io("copy file", destination.to_path_buf(), error)
-            })?;
-            written = written.saturating_add(index - start);
-        }
-    }
-
-    Ok(written)
-}
-
-pub(crate) fn destination_is_newer(source: &fs::Metadata, destination: &fs::Metadata) -> bool {
-    match (source.modified(), destination.modified()) {
-        (Ok(src), Ok(dst)) => dst > src,
-        _ => false,
-    }
-}
-
-pub(crate) fn build_delta_signature(
-    destination: &Path,
-    metadata: &fs::Metadata,
-) -> Result<Option<DeltaSignatureIndex>, LocalCopyError> {
-    let length = metadata.len();
-    if length == 0 {
-        return Ok(None);
-    }
-
-    let checksum_len = NonZeroU8::new(16).expect("strong checksum length must be non-zero");
-    let params = SignatureLayoutParams::new(length, None, ProtocolVersion::NEWEST, checksum_len);
-    let layout = match calculate_signature_layout(params) {
-        Ok(layout) => layout,
-        Err(_) => return Ok(None),
-    };
-
-    let signature = match generate_file_signature(
-        fs::File::open(destination).map_err(|error| {
-            LocalCopyError::io(
-                "read existing destination",
-                destination.to_path_buf(),
-                error,
-            )
-        })?,
-        layout,
-        SignatureAlgorithm::Md4,
-    ) {
-        Ok(signature) => signature,
-        Err(SignatureError::Io(error)) => {
-            return Err(LocalCopyError::io(
-                "read existing destination",
-                destination.to_path_buf(),
-                error,
-            ));
-        }
-        Err(_) => return Ok(None),
-    };
-
-    match DeltaSignatureIndex::from_signature(&signature, SignatureAlgorithm::Md4) {
-        Some(index) => Ok(Some(index)),
-        None => Ok(None),
-    }
-}
-
-pub(crate) struct CopyComparison<'a> {
-    pub(crate) source_path: &'a Path,
-    pub(crate) source: &'a fs::Metadata,
-    pub(crate) destination_path: &'a Path,
-    pub(crate) destination: &'a fs::Metadata,
-    pub(crate) options: &'a MetadataOptions,
-    pub(crate) size_only: bool,
-    pub(crate) checksum: bool,
-    pub(crate) checksum_algorithm: SignatureAlgorithm,
-    pub(crate) modify_window: Duration,
-}
-
-pub(crate) fn should_skip_copy(params: CopyComparison<'_>) -> bool {
-    let CopyComparison {
-        source_path,
-        source,
-        destination_path,
-        destination,
-        options,
-        size_only,
-        checksum,
-        checksum_algorithm,
-        modify_window,
-    } = params;
-    if destination.len() != source.len() {
-        return false;
-    }
-
-    if checksum {
-        return files_checksum_match(source_path, destination_path, checksum_algorithm)
-            .unwrap_or(false);
-    }
-
-    if size_only {
-        return true;
-    }
-
-    if options.times() {
-        match (source.modified(), destination.modified()) {
-            (Ok(src), Ok(dst)) if system_time_within_window(src, dst, modify_window) => {}
-            _ => return false,
-        }
-    } else {
-        return false;
-    }
-
-    files_match(source_path, destination_path)
-}
-
-pub(crate) fn system_time_within_window(a: SystemTime, b: SystemTime, window: Duration) -> bool {
-    if window.is_zero() {
-        return a.eq(&b);
-    }
-
-    match a.duration_since(b) {
-        Ok(diff) => diff <= window,
-        Err(_) => matches!(b.duration_since(a), Ok(diff) if diff <= window),
-    }
-}
-
-pub(crate) fn files_match(source: &Path, destination: &Path) -> bool {
-    let mut source_file = match fs::File::open(source) {
-        Ok(file) => file,
-        Err(_) => return false,
-    };
-    let mut destination_file = match fs::File::open(destination) {
-        Ok(file) => file,
-        Err(_) => return false,
-    };
-
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    loop {
-        let source_read = match source_file.read(&mut source_buffer) {
-            Ok(bytes) => bytes,
-            Err(_) => return false,
-        };
-        let destination_read = match destination_file.read(&mut destination_buffer) {
-            Ok(bytes) => bytes,
-            Err(_) => return false,
-        };
-
-        if source_read != destination_read {
-            return false;
-        }
-
-        if source_read == 0 {
-            return true;
-        }
-
-        if source_buffer[..source_read] != destination_buffer[..destination_read] {
-            return false;
-        }
-    }
-}
-
-pub(crate) enum StrongHasher {
-    Md4(Md4),
-    Md5(Md5),
-    Xxh64(Xxh64),
-    Xxh3(Xxh3),
-    Xxh128(Xxh3_128),
-}
-
-impl StrongHasher {
-    fn new(algorithm: SignatureAlgorithm) -> Self {
-        match algorithm {
-            SignatureAlgorithm::Md4 => StrongHasher::Md4(Md4::new()),
-            SignatureAlgorithm::Md5 => StrongHasher::Md5(Md5::new()),
-            SignatureAlgorithm::Xxh64 { seed } => StrongHasher::Xxh64(Xxh64::new(seed)),
-            SignatureAlgorithm::Xxh3 { seed } => StrongHasher::Xxh3(Xxh3::new(seed)),
-            SignatureAlgorithm::Xxh3_128 { seed } => StrongHasher::Xxh128(Xxh3_128::new(seed)),
-        }
-    }
-
-    fn update(&mut self, data: &[u8]) {
-        match self {
-            StrongHasher::Md4(state) => state.update(data),
-            StrongHasher::Md5(state) => state.update(data),
-            StrongHasher::Xxh64(state) => state.update(data),
-            StrongHasher::Xxh3(state) => state.update(data),
-            StrongHasher::Xxh128(state) => state.update(data),
-        }
-    }
-
-    fn finalize(self) -> Vec<u8> {
-        match self {
-            StrongHasher::Md4(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Md5(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh64(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh3(state) => state.finalize().as_ref().to_vec(),
-            StrongHasher::Xxh128(state) => state.finalize().as_ref().to_vec(),
-        }
-    }
-}
-
-pub(crate) fn files_checksum_match(
-    source: &Path,
-    destination: &Path,
-    algorithm: SignatureAlgorithm,
-) -> io::Result<bool> {
-    let mut source_file = fs::File::open(source)?;
-    let mut destination_file = fs::File::open(destination)?;
-
-    let mut source_hasher = StrongHasher::new(algorithm);
-    let mut destination_hasher = StrongHasher::new(algorithm);
-
-    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
-    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
-
-    loop {
-        let source_read = source_file.read(&mut source_buffer)?;
-        let destination_read = destination_file.read(&mut destination_buffer)?;
-
-        if source_read != destination_read {
-            return Ok(false);
-        }
-
-        if source_read == 0 {
-            break;
-        }
-
-        source_hasher.update(&source_buffer[..source_read]);
-        destination_hasher.update(&destination_buffer[..destination_read]);
-    }
-
-    Ok(source_hasher.finalize() == destination_hasher.finalize())
-}
