+++ crates/engine/src/local_copy/executor/file/append.rs
@@ -1,0 +1,104 @@
+use std::fs;
+use std::io::{Read, Seek, SeekFrom};
+use std::path::Path;
+
+use crate::local_copy::LocalCopyError;
+
+use super::super::super::COPY_BUFFER_SIZE;
+
+pub(crate) enum AppendMode {
+    Disabled,
+    Append(u64),
+}
+
+pub(crate) fn determine_append_mode(
+    append_allowed: bool,
+    append_verify: bool,
+    reader: &mut fs::File,
+    source: &Path,
+    destination: &Path,
+    existing_metadata: Option<&fs::Metadata>,
+    file_size: u64,
+) -> Result<AppendMode, LocalCopyError> {
+    if !append_allowed {
+        return Ok(AppendMode::Disabled);
+    }
+
+    let existing = match existing_metadata {
+        Some(meta) if meta.is_file() => meta,
+        _ => return Ok(AppendMode::Disabled),
+    };
+
+    let existing_len = existing.len();
+    if existing_len == 0 || existing_len >= file_size {
+        reader
+            .seek(SeekFrom::Start(0))
+            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
+        return Ok(AppendMode::Disabled);
+    }
+
+    if append_verify {
+        let matches = verify_append_prefix(reader, source, destination, existing_len)?;
+        reader
+            .seek(SeekFrom::Start(0))
+            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
+        if !matches {
+            return Ok(AppendMode::Disabled);
+        }
+    } else {
+        reader
+            .seek(SeekFrom::Start(0))
+            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
+    }
+
+    Ok(AppendMode::Append(existing_len))
+}
+
+fn verify_append_prefix(
+    reader: &mut fs::File,
+    source: &Path,
+    destination: &Path,
+    existing_len: u64,
+) -> Result<bool, LocalCopyError> {
+    reader
+        .seek(SeekFrom::Start(0))
+        .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
+    let mut destination_file = fs::File::open(destination).map_err(|error| {
+        LocalCopyError::io(
+            "read existing destination",
+            destination.to_path_buf(),
+            error,
+        )
+    })?;
+    let mut remaining = existing_len;
+    let mut source_buffer = vec![0u8; COPY_BUFFER_SIZE];
+    let mut destination_buffer = vec![0u8; COPY_BUFFER_SIZE];
+
+    while remaining > 0 {
+        let chunk = remaining.min(COPY_BUFFER_SIZE as u64) as usize;
+        let source_read = reader
+            .read(&mut source_buffer[..chunk])
+            .map_err(|error| LocalCopyError::io("copy file", source.to_path_buf(), error))?;
+        let destination_read = destination_file
+            .read(&mut destination_buffer[..chunk])
+            .map_err(|error| {
+                LocalCopyError::io(
+                    "read existing destination",
+                    destination.to_path_buf(),
+                    error,
+                )
+            })?;
+
+        if source_read == 0 || destination_read == 0 || source_read != destination_read {
+            return Ok(false);
+        }
+
+        if source_buffer[..source_read] != destination_buffer[..destination_read] {
+            return Ok(false);
+        }
+
+        remaining = remaining.saturating_sub(source_read as u64);
+    }
+
+    Ok(true)
+}
