+++ crates/engine/src/local_copy/executor/file/preallocate.rs
@@ -1,0 +1,76 @@
+use std::fs;
+use std::io;
+use std::path::Path;
+
+#[cfg(unix)]
+use rustix::{
+    fd::AsFd,
+    fs::{FallocateFlags, fallocate},
+    io::Errno,
+};
+
+use crate::local_copy::LocalCopyError;
+
+pub(crate) fn maybe_preallocate_destination(
+    file: &mut fs::File,
+    path: &Path,
+    total_len: u64,
+    existing_bytes: u64,
+    enabled: bool,
+) -> Result<(), LocalCopyError> {
+    if !enabled || total_len == 0 || total_len <= existing_bytes {
+        return Ok(());
+    }
+
+    preallocate_destination_file(file, path, total_len)
+}
+
+fn preallocate_destination_file(
+    file: &mut fs::File,
+    path: &Path,
+    total_len: u64,
+) -> Result<(), LocalCopyError> {
+    #[cfg(unix)]
+    {
+        if total_len == 0 {
+            return Ok(());
+        }
+
+        if total_len > i64::MAX as u64 {
+            return Err(LocalCopyError::io(
+                "preallocate destination file",
+                path.to_path_buf(),
+                io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "preallocation size exceeds platform limit",
+                ),
+            ));
+        }
+
+        let fd = file.as_fd();
+        match fallocate(fd, FallocateFlags::empty(), 0, total_len) {
+            Ok(()) => Ok(()),
+            Err(Errno::OPNOTSUPP | Errno::NOSYS | Errno::INVAL) => {
+                file.set_len(total_len).map_err(|error| {
+                    LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
+                })
+            }
+            Err(errno) => Err(LocalCopyError::io(
+                "preallocate destination file",
+                path.to_path_buf(),
+                io::Error::from_raw_os_error(errno.raw_os_error()),
+            )),
+        }
+    }
+
+    #[cfg(not(unix))]
+    {
+        if total_len == 0 {
+            return Ok(());
+        }
+
+        file.set_len(total_len).map_err(|error| {
+            LocalCopyError::io("preallocate destination file", path.to_path_buf(), error)
+        })
+    }
+}
