+++ crates/engine/src/local_copy/executor/file/sparse.rs
@@ -1,0 +1,46 @@
+use std::fs;
+use std::io::{Seek, SeekFrom, Write};
+use std::path::Path;
+
+use crate::local_copy::LocalCopyError;
+
+pub(crate) fn write_sparse_chunk(
+    writer: &mut fs::File,
+    chunk: &[u8],
+    destination: &Path,
+) -> Result<usize, LocalCopyError> {
+    let mut index = 0usize;
+    let mut written = 0usize;
+
+    while index < chunk.len() {
+        if chunk[index] == 0 {
+            let start = index;
+            while index < chunk.len() && chunk[index] == 0 {
+                index += 1;
+            }
+            let span = index - start;
+            if span > 0 {
+                writer
+                    .seek(SeekFrom::Current(span as i64))
+                    .map_err(|error| {
+                        LocalCopyError::io(
+                            "seek in destination file",
+                            destination.to_path_buf(),
+                            error,
+                        )
+                    })?;
+            }
+        } else {
+            let start = index;
+            while index < chunk.len() && chunk[index] != 0 {
+                index += 1;
+            }
+            writer.write_all(&chunk[start..index]).map_err(|error| {
+                LocalCopyError::io("copy file", destination.to_path_buf(), error)
+            })?;
+            written = written.saturating_add(index - start);
+        }
+    }
+
+    Ok(written)
+}
